// src/pages/ProgramDiscBasic/types/types.ts
export interface IAddCompetenceValues {
	name: string;
	code: string;
}

export interface IEditCompetenceValues {
	name: string;
	code: string;
}

export interface IAddDisciplineValues {
	name: string;
	area: string;
	description: string;
}

export interface IEditDisciplineValues {
	name: string;
	area: string;
	description: string;
}